perf(flash): replace double-precision math with float / inline cubing
Two changes that drop 3048 bytes of flash (127492 → 124444) with
identical runtime behavior:

1. animation_engine.h: sin() → sinf() in the WAVE phase calculation.
   The implicit double promotion was pulling in __kernel_cos,
   __kernel_rem_pio2, __ieee754_rem_pio2, plus the double-precision
   add/sub/mul/div helpers — ~3 KB combined — even though the result
   was immediately cast back to float.

2. dc_motor.h: pow(u, 3.0f) → u*u*u in the ease-in-out-cubic ramp.
   Avoids powf (~936 bytes) and its remaining double-precision
   dependencies. Hand-expanding x^3 produces identical output and is
   also faster.

Step 1 toward servo-dz7 (need to free 4612 bytes total to fit OTA's
122880-byte partition; this lands 3048 of them). All host tests still
pass: 7/7 time, 6/6 motion, 22/22 storage.

+++ adafruit_16_servo/animation_engine.h
@@ -47,7 +47,7 @@ void updateWave() {
     float phase = (float)elapsed / cycleMs;
     phase += (float)(servoIndex * wavePhaseOffset) / 360.0f;
 
-    float sineVal = sin(phase * 2.0f * PI);
+    float sineVal = sinf(phase * 2.0f * PI);
     float centerDegrees = ((float)minDegrees + (float)maxDegrees) / 2.0f;
     float degrees = centerDegrees + (sineVal * (float)waveAmplitude / 2.0f);
     uint16_t pulse = degreesToPulse(i, (uint16_t)constrain(degrees, minDegrees, maxDegrees));
